nvim: Check that `environ` exists before using it
It seems that, under some circumstances (I'm not sure, but starting nvim
from a bash script seems to be one of them), `environ` doesn't (yet?)
exist and nvim starts up with a nasty error message complaining about
that.

Then again, if `environ` doesn't exist, we're (probably) not going to
need CoC or its mappings either. So just check if it exists before
loading either.