nvim: Remove `let g:clojure_align_subforms = 1`
Re-reading the Clojure Style Guide (with its recent changes), I agree
with the position with "broad community support" that a simpler policy
is better: just use 2 spaces.

Given that I'm not an old school lisper, and none of the code bases I
work in is strongly lisp-flavoured, this seems like the more pragmatic
position.