nvim: Replace Avante.nvim with CodeCompanion
After lots of searching and tinkering, the best I could get out of
Avante is to work with Ollama using CPU inference, presumably because
Avante's system prompt blows the context/VRAM limit. CodeCompanion OTOH
worked with GPU inference on the first try.

Avante may be a bit more polished, but that doesn't help if I can't use
it with Ollama.